Belt sanding compute budget and other miscellany
- vm_compute_budget replaced with faster and more secure compile time
  macros and added unit-test coverage.  If we need to change individual
  items at run-time (e.g.  feature flag enabled with a different compute
  budget), those affected variables can be isolated as run time
  variables while both need to be present in the code (or we can flip
  back to run-time).

- Made fd_vm_t entrypoint ulong to match the vast majority of usage.
  (The sole remaining usage is suspect but important so needs to be
  investigated further.)

- Made fd_vm_t syscalls and calldests const and uses the nifty new const
  correct map query to make fd_vm_t syscalls queries const correct.

- Some more reorg of the layout of the fd_vm_t and additional commentary
  for future improvements.

- Minor cleanup to fd_vm_trace internals to be consistent with other
  implementations.

- Eliminated the fd_vm_t dependencies in fd_vm_tool disasm and added
  fixmes for potential future directions for improving disasm
  diagnostics.

- Plumbed through changes to all users.

Passes the usual unit tests

+++ src/flamenco/runtime/program/fd_bpf_loader_v3_program.c
@@ -166,7 +166,7 @@ fd_bpf_loader_v3_user_execute( fd_exec_instr_ctx_t ctx ) {
 
   // FD_LOG_DEBUG(("Starting CUs %lu", ctx.txn_ctx->compute_meter));
   fd_vm_t vm = {
-    .entrypoint                 = (long)prog->entry_pc,
+    .entrypoint                 = prog->entry_pc,
     .syscalls                   = syscalls,
     .calldests                  = prog->calldests,
     .program_counter            = 0,
@@ -288,7 +288,7 @@ setup_program( fd_exec_instr_ctx_t * ctx,
   }
 
   fd_vm_t vm = {
-    .entrypoint          = (long)prog->entry_pc,
+    .entrypoint          = prog->entry_pc,
     .syscalls            = syscalls,
     .calldests           = prog->calldests,
     .program_counter     = 0,
